Добален метод получения данных курса валют с сайта через делегат

+++ Delivery Calculator/DecodableModel.swift
@@ -18,7 +18,7 @@ struct ValuteAnswer: Decodable {
     var ID: String = ""
     var NumCode: String = ""
     var CharCode: String = ""
-    var Nominal: Double = 0
+    var Nominal: Int = 0
     var Name: String = ""
     var Value: Double = 0
     var Previous: Double = 0
